+++ src/adapter/graphql/src/queries/accounts/accounts.rs
@@ -19,8 +19,6 @@ pub struct Accounts;
 #[Object]
 impl Accounts {
     /// Returns account by its ID
-    #[allow(unused_variables)]
-    #[allow(clippy::unused_async)]
     async fn by_id(&self, ctx: &Context<'_>, account_id: AccountID) -> Result<Option<Account>> {
         let authentication_service =
             from_catalog::<dyn kamu_accounts::AuthenticationService>(ctx).unwrap();