+++ src/domain/accounts/domain/src/entities/account.rs
@@ -39,7 +39,7 @@ pub struct Account {
     pub account_type: AccountType,
     pub avatar_url: Option<String>,
     pub registered_at: DateTime<Utc>,
-    // TODO: ReBAC: absorb the `is_admin` attribute from the Accounts domain
+    // TODO: Private Datasets: absorb the `is_admin` attribute from the Accounts domain
     //       https://github.com/kamu-data/kamu-cli/issues/766
     pub is_admin: bool,
     pub provider: String,