+++ src/domain/accounts/domain/src/services/account_service.rs
@@ -1,0 +1,40 @@
+// Copyright Kamu Data, Inc. and contributors. All rights reserved.
+//
+// Use of this software is governed by the Business Source License
+// included in the LICENSE file.
+//
+// As of the Change Date specified in that file, in accordance with
+// the Business Source License, use of this software will be governed
+// by the Apache License, Version 2.0.
+
+use database_common::{EntityPageListing, PaginationOpts};
+use internal_error::InternalError;
+use thiserror::Error;
+
+use crate::{Account, AccountPageStream};
+
+////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
+
+// TODO: Private Datasets: tests
+#[async_trait::async_trait]
+pub trait AccountService: Sync + Send {
+    // TODO: Private Datasets: extract to AccountRegistry?
+    fn all_accounts(&self) -> AccountPageStream;
+
+    async fn list_all_accounts(
+        &self,
+        pagination: PaginationOpts,
+    ) -> Result<EntityPageListing<Account>, ListAccountError>;
+}
+
+////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
+// Error
+////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
+
+#[derive(Debug, Error)]
+pub enum ListAccountError {
+    #[error(transparent)]
+    Internal(#[from] InternalError),
+}
+
+////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////