+++ src/domain/auth-rebac/services/src/rebac_service_impl.rs
@@ -7,12 +7,16 @@
 // the Business Source License, use of this software will be governed
 // by the Apache License, Version 2.0.
 
+use std::collections::HashMap;
 use std::sync::Arc;
 
 use dill::{component, interface};
+use internal_error::{ErrorIntoInternal, ResultIntoInternal};
 use kamu_auth_rebac::{
+    AccountProperties,
     AccountPropertyName,
     AccountToDatasetRelation,
+    DatasetProperties,
     DatasetPropertyName,
     DeleteEntitiesRelationError,
     DeleteEntityPropertiesError,
@@ -21,7 +25,7 @@ use kamu_auth_rebac::{
     DeleteRelationError,
     Entity,
     EntityWithRelation,
-    GetEntityPropertiesError,
+    GetPropertiesError,
     InsertEntitiesRelationError,
     InsertRelationError,
     PropertyName,
@@ -32,8 +36,9 @@ use kamu_auth_rebac::{
     SetEntityPropertyError,
     SubjectEntityRelationsError,
     UnsetEntityPropertyError,
+    PROPERTY_VALUE_BOOLEAN_TRUE,
 };
-use opendatafabric::{AccountID, DatasetID};
+use opendatafabric as odf;
 
 ////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
 
@@ -53,7 +58,7 @@ impl RebacServiceImpl {
 impl RebacService for RebacServiceImpl {
     async fn set_account_property(
         &self,
-        account_id: &AccountID,
+        account_id: &odf::AccountID,
         property_name: AccountPropertyName,
         property_value: &PropertyValue,
     ) -> Result<(), SetEntityPropertyError> {
@@ -67,7 +72,7 @@ impl RebacService for RebacServiceImpl {
 
     async fn unset_account_property(
         &self,
-        account_id: &AccountID,
+        account_id: &odf::AccountID,
         property_name: AccountPropertyName,
     ) -> Result<(), UnsetEntityPropertyError> {
         use futures::FutureExt;
@@ -83,59 +88,75 @@ impl RebacService for RebacServiceImpl {
 
     async fn get_account_properties(
         &self,
-        account_id: &AccountID,
-    ) -> Result<Vec<(PropertyName, PropertyValue)>, GetEntityPropertiesError> {
+        account_id: &odf::AccountID,
+    ) -> Result<AccountProperties, GetPropertiesError> {
         let account_id = account_id.as_did_str().to_stack_string();
         let account_entity = Entity::new_account(account_id.as_str());
 
-        let properties = self
+        let entity_properties = self
             .rebac_repo
             .get_entity_properties(&account_entity)
-            .await?;
+            .await
+            .int_err()?;
 
-        Ok(properties)
+        let account_properties = entity_properties
+            .into_iter()
+            .map(|(name, value)| match name {
+                PropertyName::Dataset(_) => unreachable!(),
+                PropertyName::Account(account_property_name) => (account_property_name, value),
+            })
+            .fold(AccountProperties::default(), |mut acc, (name, value)| {
+                match name {
+                    AccountPropertyName::IsAnAdmin => {
+                        acc.is_admin = value == PROPERTY_VALUE_BOOLEAN_TRUE;
+                    }
+                };
+                acc
+            });
+
+        Ok(account_properties)
     }
 
     async fn set_dataset_property(
         &self,
-        dataset_id: &DatasetID,
+        dataset_id: &odf::DatasetID,
         property_name: DatasetPropertyName,
         property_value: &PropertyValue,
     ) -> Result<(), SetEntityPropertyError> {
         let dataset_id = dataset_id.as_did_str().to_stack_string();
-        let dataset_id_entity = Entity::new_dataset(dataset_id.as_str());
+        let dataset_entity = Entity::new_dataset(dataset_id.as_str());
 
         self.rebac_repo
-            .set_entity_property(&dataset_id_entity, property_name.into(), property_value)
+            .set_entity_property(&dataset_entity, property_name.into(), property_value)
             .await
     }
 
     async fn unset_dataset_property(
         &self,
-        dataset_id: &DatasetID,
+        dataset_id: &odf::DatasetID,
         property_name: DatasetPropertyName,
     ) -> Result<(), UnsetEntityPropertyError> {
         use futures::FutureExt;
 
         let dataset_id = dataset_id.as_did_str().to_stack_string();
-        let dataset_id_entity = Entity::new_dataset(dataset_id.as_str());
+        let dataset_entity = Entity::new_dataset(dataset_id.as_str());
 
         self.rebac_repo
-            .delete_entity_property(&dataset_id_entity, property_name.into())
+            .delete_entity_property(&dataset_entity, property_name.into())
             .map(map_delete_entity_property_result)
             .await
     }
 
     async fn delete_dataset_properties(
         &self,
-        dataset_id: &DatasetID,
+        dataset_id: &odf::DatasetID,
     ) -> Result<(), DeletePropertiesError> {
         let dataset_id = dataset_id.as_did_str().to_stack_string();
-        let dataset_id_entity = Entity::new_dataset(dataset_id.as_str());
+        let dataset_entity = Entity::new_dataset(dataset_id.as_str());
 
         match self
             .rebac_repo
-            .delete_entity_properties(&dataset_id_entity)
+            .delete_entity_properties(&dataset_entity)
             .await
         {
             Ok(_) => Ok(()),
@@ -148,24 +169,93 @@ impl RebacService for RebacServiceImpl {
 
     async fn get_dataset_properties(
         &self,
-        dataset_id: &DatasetID,
-    ) -> Result<Vec<(PropertyName, PropertyValue)>, GetEntityPropertiesError> {
+        dataset_id: &odf::DatasetID,
+    ) -> Result<DatasetProperties, GetPropertiesError> {
         let dataset_id = dataset_id.as_did_str().to_stack_string();
-        let dataset_id_entity = Entity::new_dataset(dataset_id.as_str());
+        let dataset_entity = Entity::new_dataset(dataset_id.as_str());
 
-        let properties = self
+        let entity_properties = self
             .rebac_repo
-            .get_entity_properties(&dataset_id_entity)
-            .await?;
+            .get_entity_properties(&dataset_entity)
+            .await
+            .int_err()?;
 
-        Ok(properties)
+        let dataset_properties = entity_properties
+            .into_iter()
+            .map(|(name, value)| match name {
+                PropertyName::Dataset(dataset_property_name) => (dataset_property_name, value),
+                PropertyName::Account(_) => unreachable!(),
+            })
+            .fold(DatasetProperties::default(), |mut acc, (name, value)| {
+                match name {
+                    DatasetPropertyName::AllowsAnonymousRead => {
+                        acc.allows_anonymous_read = value == PROPERTY_VALUE_BOOLEAN_TRUE;
+                    }
+                    DatasetPropertyName::AllowsPublicRead => {
+                        acc.allows_public_read = value == PROPERTY_VALUE_BOOLEAN_TRUE;
+                    }
+                };
+                acc
+            });
+
+        Ok(dataset_properties)
+    }
+
+    async fn get_dataset_properties_by_ids(
+        &self,
+        dataset_ids: &[odf::DatasetID],
+    ) -> Result<HashMap<odf::DatasetID, DatasetProperties>, GetPropertiesError> {
+        let dataset_entities = dataset_ids
+            .iter()
+            .map(|id| Entity::new_dataset(id.to_string()))
+            .collect::<Vec<_>>();
+
+        let entity_properties = self
+            .rebac_repo
+            .get_entity_properties_by_ids(&dataset_entities)
+            .await
+            .int_err()?;
+
+        let mut dataset_properties_map = HashMap::new();
+
+        for dataset_id in dataset_ids {
+            dataset_properties_map.insert(dataset_id.clone(), DatasetProperties::default());
+        }
+
+        let entity_properties_it =
+            entity_properties
+                .into_iter()
+                .map(|(entity, name, value)| match name {
+                    PropertyName::Dataset(dataset_property_name) => {
+                        (entity.entity_id, dataset_property_name, value)
+                    }
+                    PropertyName::Account(_) => unreachable!(),
+                });
+
+        for (entity_id, name, value) in entity_properties_it {
+            let dataset_id = odf::DatasetID::from_did_str(&entity_id).int_err()?;
+            let dataset_properties = dataset_properties_map
+                .get_mut(&dataset_id)
+                .ok_or_else(|| format!("dataset_id not found: {dataset_id}").int_err())?;
+
+            match name {
+                DatasetPropertyName::AllowsAnonymousRead => {
+                    dataset_properties.allows_anonymous_read = value == PROPERTY_VALUE_BOOLEAN_TRUE;
+                }
+                DatasetPropertyName::AllowsPublicRead => {
+                    dataset_properties.allows_public_read = value == PROPERTY_VALUE_BOOLEAN_TRUE;
+                }
+            };
+        }
+
+        Ok(dataset_properties_map)
     }
 
     async fn insert_account_dataset_relation(
         &self,
-        account_id: &AccountID,
+        account_id: &odf::AccountID,
         relationship: AccountToDatasetRelation,
-        dataset_id: &DatasetID,
+        dataset_id: &odf::DatasetID,
     ) -> Result<(), InsertRelationError> {
         use futures::FutureExt;
 
@@ -173,13 +263,13 @@ impl RebacService for RebacServiceImpl {
         let account_entity = Entity::new_account(account_id.as_str());
 
         let dataset_id = dataset_id.as_did_str().to_stack_string();
-        let dataset_id_entity = Entity::new_dataset(dataset_id.as_str());
+        let dataset_entity = Entity::new_dataset(dataset_id.as_str());
 
         self.rebac_repo
             .insert_entities_relation(
                 &account_entity,
                 Relation::AccountToDataset(relationship),
-                &dataset_id_entity,
+                &dataset_entity,
             )
             .map(|res| match res {
                 Ok(_) => Ok(()),
@@ -195,22 +285,22 @@ impl RebacService for RebacServiceImpl {
 
     async fn delete_account_dataset_relation(
         &self,
-        account_id: &AccountID,
+        account_id: &odf::AccountID,
         relationship: AccountToDatasetRelation,
-        dataset_id: &DatasetID,
+        dataset_id: &odf::DatasetID,
     ) -> Result<(), DeleteRelationError> {
         let account_id = account_id.as_did_str().to_stack_string();
         let account_entity = Entity::new_account(account_id.as_str());
 
         let dataset_id = dataset_id.as_did_str().to_stack_string();
-        let dataset_id_entity = Entity::new_dataset(dataset_id.as_str());
+        let dataset_entity = Entity::new_dataset(dataset_id.as_str());
 
         match self
             .rebac_repo
             .delete_entities_relation(
                 &account_entity,
                 Relation::AccountToDataset(relationship),
-                &dataset_id_entity,
+                &dataset_entity,
             )
             .await
         {
@@ -224,7 +314,7 @@ impl RebacService for RebacServiceImpl {
 
     async fn get_account_dataset_relations(
         &self,
-        account_id: &AccountID,
+        account_id: &odf::AccountID,
     ) -> Result<Vec<EntityWithRelation>, SubjectEntityRelationsError> {
         let account_id = account_id.as_did_str().to_stack_string();
         let account_entity = Entity::new_account(account_id.as_str());