+++ src/domain/core/src/auth/dataset_action_authorizer.rs
@@ -12,7 +12,7 @@ use std::str::FromStr;
 
 use dill::*;
 use internal_error::{ErrorIntoInternal, InternalError};
-use opendatafabric::{DatasetHandle, DatasetRef};
+use opendatafabric as odf;
 use thiserror::Error;
 
 use crate::AccessError;
@@ -23,13 +23,14 @@ use crate::AccessError;
 pub trait DatasetActionAuthorizer: Sync + Send {
     async fn check_action_allowed(
         &self,
-        dataset_handle: &DatasetHandle,
+        // TODO: Private Datasets: use odf::DatasetID, here and below
+        dataset_handle: &odf::DatasetHandle,
         action: DatasetAction,
     ) -> Result<(), DatasetActionUnauthorizedError>;
 
     async fn is_action_allowed(
         &self,
-        dataset_handle: &DatasetHandle,
+        dataset_handle: &odf::DatasetHandle,
         action: DatasetAction,
     ) -> Result<bool, InternalError> {
         match self.check_action_allowed(dataset_handle, action).await {
@@ -39,17 +40,24 @@ pub trait DatasetActionAuthorizer: Sync + Send {
         }
     }
 
-    async fn get_allowed_actions(&self, dataset_handle: &DatasetHandle) -> HashSet<DatasetAction>;
+    // TODO: Private Datasets: tests
+    async fn get_allowed_actions(
+        &self,
+        dataset_handle: &odf::DatasetHandle,
+    ) -> Result<HashSet<DatasetAction>, InternalError>;
 
+    // TODO: Private Datasets: tests
     async fn filter_datasets_allowing(
         &self,
-        dataset_handles: Vec<DatasetHandle>,
+        // TODO: Private Datasets: use slice? here and above
+        dataset_handles: Vec<odf::DatasetHandle>,
         action: DatasetAction,
-    ) -> Result<Vec<DatasetHandle>, InternalError>;
+    ) -> Result<Vec<odf::DatasetHandle>, InternalError>;
 
+    // TODO: Private Datasets: tests
     async fn classify_datasets_by_allowance(
         &self,
-        dataset_handles: Vec<DatasetHandle>,
+        dataset_handles: Vec<odf::DatasetHandle>,
         action: DatasetAction,
     ) -> Result<ClassifyByAllowanceResponse, InternalError>;
 }
@@ -85,6 +93,29 @@ impl std::fmt::Display for DatasetAction {
     }
 }
 
+#[cfg(feature = "oso")]
+impl oso::FromPolar for DatasetAction {
+    fn from_polar(polar_value: oso::PolarValue) -> oso::Result<Self> {
+        use oso::errors::{OsoError, TypeError};
+        use oso::PolarValue;
+
+        let PolarValue::String(raw_dataset_action) = polar_value else {
+            return Err(TypeError::expected("String").user());
+        };
+
+        Self::from_str(&raw_dataset_action).map_err(|e| OsoError::Custom {
+            message: e.to_string(),
+        })
+    }
+}
+
+#[cfg(feature = "oso")]
+impl oso::ToPolar for DatasetAction {
+    fn to_polar(self) -> oso::PolarValue {
+        self.to_string().to_polar()
+    }
+}
+
 ////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
 
 #[derive(Debug, Error)]
@@ -93,22 +124,22 @@ pub enum DatasetActionUnauthorizedError {
     Access(AccessError),
 
     #[error(transparent)]
-    Internal(InternalError),
+    Internal(#[from] InternalError),
 }
 
 #[derive(Debug, Error)]
 #[error("User has no '{action}' permission in dataset '{dataset_ref}'")]
 pub struct DatasetActionNotEnoughPermissionsError {
     pub action: DatasetAction,
-    pub dataset_ref: DatasetRef,
+    pub dataset_ref: odf::DatasetRef,
 }
 
 ////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
 
 #[derive(Debug)]
 pub struct ClassifyByAllowanceResponse {
-    pub authorized_handles: Vec<DatasetHandle>,
-    pub unauthorized_handles_with_errors: Vec<(DatasetHandle, DatasetActionUnauthorizedError)>,
+    pub authorized_handles: Vec<odf::DatasetHandle>,
+    pub unauthorized_handles_with_errors: Vec<(odf::DatasetHandle, DatasetActionUnauthorizedError)>,
 }
 
 ////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
@@ -127,28 +158,31 @@ impl AlwaysHappyDatasetActionAuthorizer {
 impl DatasetActionAuthorizer for AlwaysHappyDatasetActionAuthorizer {
     async fn check_action_allowed(
         &self,
-        _dataset_handle: &DatasetHandle,
+        _dataset_handle: &odf::DatasetHandle,
         _action: DatasetAction,
     ) -> Result<(), DatasetActionUnauthorizedError> {
         // Ignore rules
         Ok(())
     }
 
-    async fn get_allowed_actions(&self, _dataset_handle: &DatasetHandle) -> HashSet<DatasetAction> {
-        HashSet::from([DatasetAction::Read, DatasetAction::Write])
+    async fn get_allowed_actions(
+        &self,
+        _dataset_handle: &odf::DatasetHandle,
+    ) -> Result<HashSet<DatasetAction>, InternalError> {
+        Ok(HashSet::from([DatasetAction::Read, DatasetAction::Write]))
     }
 
     async fn filter_datasets_allowing(
         &self,
-        dataset_handles: Vec<DatasetHandle>,
+        dataset_handles: Vec<odf::DatasetHandle>,
         _action: DatasetAction,
-    ) -> Result<Vec<DatasetHandle>, InternalError> {
+    ) -> Result<Vec<odf::DatasetHandle>, InternalError> {
         Ok(dataset_handles)
     }
 
     async fn classify_datasets_by_allowance(
         &self,
-        dataset_handles: Vec<DatasetHandle>,
+        dataset_handles: Vec<odf::DatasetHandle>,
         _action: DatasetAction,
     ) -> Result<ClassifyByAllowanceResponse, InternalError> {
         Ok(ClassifyByAllowanceResponse {