+++ src/infra/core/tests/tests/test_dependency_graph_inmem.rs
@@ -776,7 +776,6 @@ impl DependencyGraphHarness {
             .dependency_graph_service
             .get_recursive_upstream_dependencies(dataset_ids)
             .await
-            .int_err()
             .unwrap()
             .collect()
             .await;